Initial commit - Deploy OpenStack Pike with Kolla
This code sets up an underlying Ubuntu environment in order
to deploy a Pike based OpenStack.  Note that this does not
currently automatically build a set of Pike (5.0.0) basd images
and does not configure a docker registry to store them, as this
is left as an exercies for the user.